requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
langchain-community>=0.0.10
llama-cpp-python>=0.2.0
python-dotenv>=0.19.0 